requirements.txt
Django==5.0.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.1
Pillow==11.3.0
PyJWT==2.10.1